# gin-t/docs/swagger.yaml
basePath: /api/v1
definitions:
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  model.Member:
    properties:
      addip:
        type: string
      addtime:
        type: string
      area:
        type: integer
      area_cn:
        type: string
      attention:
        type: integer
      avatar:
        type: string
      bid:
        type: integer
      city:
        type: integer
      city_cn:
        type: string
      gid:
        type: integer
      id:
        type: integer
      jointime:
        type: string
      loginip:
        type: string
      loginnum:
        type: integer
      logintime:
        type: string
      mobile:
        type: string
      monicker:
        type: string
      name:
        type: string
      number:
        type: string
      openid_app:
        type: string
      openid_miniapp:
        type: string
      openid_public:
        type: string
      openid_web:
        type: string
      order:
        type: string
      password:
        type: string
      position:
        type: integer
      province:
        type: integer
      province_cn:
        type: string
      salt:
        type: string
      sex:
        type: integer
      unionid:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: gin测试api
  version: "1.0"
paths:
  /dept/getdeptone:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: Account ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/model.Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: 用户信息
      tags:
      - Dept
  /member/account:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: Account ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          headers:
            Token:
              description: qwerty
              type: string
          schema:
            $ref: '#/definitions/model.Member'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: 用户信息
      tags:
      - Member
swagger: "2.0"